Add daemon and proper logging support
The -D flag now causes mongrel to daemonize. It was previously a no op. This creates a log file (mongrel.log) and a pid file (mongrel.pid) in the log directory (<Park Place directory>/log/). Moved the camping.log file, created when debug is set, to that directory as well.